feat: add download transcript button
The "download transcripts" button already exists, though it actually
downloads the captions not transcripts. Hence that button -- and the
related code -- is renamed to use the term "captions". Because the
original intention was to add a download transcripts button in #6, this
commit adds the originally required button and functionality as well.

Besides that, it worth mentioning the transcripts are different in a way
compared to captions in a way that it is not bind to any language.